requirements.txt
fastapi
uvicorn
asyncpg
Jinja2
python-dotenv
celery
redis
googletrans
aiohttp
requests
pydantic-settings
python-jose